Rename TreehouseLazyColumn to LazyListIntervalContent
